backend/src/generate_embeddings/requirements.txt
boto3
botocore
aws-lambda-powertools
langchain-community
langchain-openai
langchain-core
langchain
faiss-cpu
pypdf
urllib3
PyMuPDF==1.22.5 
Pillow==10.0.0   
pytesseract==0.3.10  
